use-case-2-data-extraction-and-analysis/requirements.txt
prefect==2.20.3
psycopg2-binary==2.9.9
requests==2.32.3
